+++ canvasapi/lti_resource_link.py
@@ -1,0 +1,6 @@
+from canvasapi.canvas_object import CanvasObject
+
+
+class LTIResourceLink(CanvasObject):
+    def __str__(self):
+        return "{} ({})".format(self.url, self.title)