+++ tests/test_course.py
@@ -35,6 +35,7 @@ from canvasapi.grading_period import GradingPeriod
 from canvasapi.grading_standard import GradingStandard
 from canvasapi.group import Group, GroupCategory
 from canvasapi.license import License
+from canvasapi.lti_resource_link import LTIResourceLink
 from canvasapi.module import Module
 from canvasapi.new_quiz import NewQuiz
 from canvasapi.outcome import OutcomeGroup, OutcomeLink, OutcomeResult
@@ -1889,6 +1890,48 @@ class TestCourse(unittest.TestCase):
         self.assertEqual(len(root_folder_list), 1)
         self.assertIsInstance(root_folder_list[0], Folder)
         self.assertEqual("course_files", root_folder_list[0].name)
+
+    # create_lti_resource_link()
+    def test_create_lti_resource_link(self, m):
+        register_uris({"lti_resource_link": ["create_lti_resource_link"]}, m)
+        custom_dict = {"hello": "world"}
+
+        evnt = self.course.create_lti_resource_link(
+            url="https://example.com/lti/launch/content_item/123",
+            title="Test LTI Resource Link",
+            custom=custom_dict,
+        )
+        self.assertIsInstance(evnt, LTIResourceLink)
+
+        self.assertEqual(evnt.title, "Test LTI Resource Link")
+        self.assertEqual(evnt.url, "https://example.com/lti/launch/content_item/123")
+
+    def test_create_lti_resource_link_fail(self, m):
+        with self.assertRaises(RequiredFieldMissing):
+            self.course.create_lti_resource_link({})
+
+    # get_lti_resource_links()
+    def test_get_lti_resource_links(self, m):
+        register_uris({"lti_resource_link": ["list_lti_resource_links"]}, m)
+
+        lti_resource_links = self.course.get_lti_resource_links()
+        lti_resource_link_list = [link for link in lti_resource_links]
+        self.assertEqual(len(lti_resource_link_list), 3)
+        self.assertIsInstance(lti_resource_link_list[0], LTIResourceLink)
+
+    # get_lti_resource_link()
+    def test_get_lti_resource_link(self, m):
+        register_uris({"lti_resource_link": ["get_lti_resource_link"]}, m)
+
+        lti_resource_link_by_id = self.course.get_lti_resource_link(45)
+
+        self.assertIsInstance(lti_resource_link_by_id, LTIResourceLink)
+        self.assertEqual(lti_resource_link_by_id.title, "Test LTI Resource Link")
+        lti_resource_link_by_obj = self.course.get_lti_resource_link(
+            lti_resource_link_by_id
+        )
+        self.assertIsInstance(lti_resource_link_by_obj, LTIResourceLink)
+        self.assertEqual(lti_resource_link_by_obj.title, "Test LTI Resource Link")
 
 
 @requests_mock.Mocker()